Separate component skeletons
