+++ src/modules/overview/components/Promo/index.ts
@@ -1,1 +1,2 @@
 export * from './Promo';
+export * from './PromoSkeleton';
