+++ src/modules/overview/components/PromoCard/index.ts
@@ -1,1 +1,2 @@
 export * from './PromoCard';
+export * from './PromoCardSkeleton';
