--- src/modules/overview/components/PromoSkeleton/index.ts
@@ -1,1 +1,0 @@
-export * from './PromoSkeleton';
