+++ src/modules/overview/components/PromoThumb/index.ts
@@ -1,1 +1,2 @@
 export * from './PromoThumb';
+export * from './PromoThumbSkeleton';
